fix(score): put freshness back on the server's 0-10 scale
The server takes and returns scores out of 10 (validated 0..10), but the whole
score UI was written for a 0-100 scale: the ring divided by 100, the colour
thresholds were 90/70, and the seller's warning threshold was 60. A 9.2/10
product therefore drew as a nearly empty ring in red -- the app told buyers
that the freshest goods were the most suspect.

Product.freshnessScore was also declared int, so 9.2 was truncated to 9 at
parse time.

- freshnessScore becomes a double, keeping what the server sent
- adds a maxScore = 10 constant for the ring to divide by
- colour thresholds 9 / 7, seller warning threshold 6
- scores render to one decimal place instead of as whole numbers
- the shop page's product card uses ProductThumbnail (it drew its own grey box
  and ignored the seller's real photo)

+++ lib/data/models/product.dart
@@ -6,7 +6,12 @@ class Product {
   String name;
   String description;
   String category;
-  int freshnessScore;
+  /// Freshness on the server's 0-10 scale.
+  ///
+  /// Was an int, which turned a 9.2 from the server into a 9 — and the score
+  /// widgets read it as if it were out of 100, so every real product rendered
+  /// red at 9%.
+  double freshnessScore;
   String freshnessNote;
   int price;
   List<String> tags;
@@ -36,7 +41,7 @@ class Product {
       name: json['name'] as String,
       description: json['description'] as String,
       category: json['category'] as String,
-      freshnessScore: (json['freshnessScore'] as num).toInt(),
+      freshnessScore: (json['freshnessScore'] as num).toDouble(),
       freshnessNote: json['freshnessNote'] as String,
       price: (json['price'] as num).toInt(),
       tags: stringList(json['tags']),
